+++ src/main/java/lucie/hitchhike/item/ItemAlias.java
@@ -11,5 +11,11 @@ public class ItemAlias
     public static Item POUCH;
 
     @ObjectHolder("pouch_with_horse")
-    public static Item POUCH_WITH_HORSE;
+    public static ItemContent POUCH_WITH_HORSE;
+
+    @ObjectHolder("pouch_with_skeleton_horse")
+    public static ItemContent POUCH_WITH_SKELETON_HORSE;
+
+    @ObjectHolder("pouch_with_zombie_horse")
+    public static ItemContent POUCH_WITH_ZOMBIE_HORSE;
 }