+++ src/main/java/lucie/hitchhike/util/UtilPouch.java
@@ -1,25 +1,179 @@
 package lucie.hitchhike.util;
 
+import lucie.hitchhike.Hitchhike;
 import lucie.hitchhike.item.ItemAlias;
+import lucie.hitchhike.item.ItemContent;
+import net.minecraft.core.particles.ParticleTypes;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.sounds.SoundEvents;
 import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.animal.horse.AbstractHorse;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.enchantment.EnchantmentHelper;
 import net.minecraft.world.phys.Vec3;
+import net.minecraftforge.event.entity.player.PlayerInteractEvent;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
+import net.minecraftforge.fml.common.Mod;
 
+import javax.annotation.Nullable;
+import java.util.Arrays;
+import java.util.List;
+
+@Mod.EventBusSubscriber(modid = Hitchhike.MODID)
 public class UtilPouch
 {
-    public static ItemStack release(AbstractHorse horse, Player player, InteractionHand hand, Vec3 pos, boolean ride)
+    /* Events */
+
+    @SubscribeEvent
+    public static void eventClick(PlayerInteractEvent.RightClickItem event)
     {
+        List<ItemContent> horses = Arrays.asList(ItemAlias.POUCH_WITH_HORSE, ItemAlias.POUCH_WITH_SKELETON_HORSE, ItemAlias.POUCH_WITH_ZOMBIE_HORSE);
+
+        // Release
+        if (event.getItemStack().getItem() instanceof ItemContent)
+        {
+            InteractionResult result = release((ItemContent) event.getItemStack().getItem(), event.getPlayer(), event.getHand(), null);
+
+            if (result.equals(InteractionResult.SUCCESS))
+            {
+                event.setCancellationResult(InteractionResult.SUCCESS);
+                event.setCanceled(true);
+            }
+        }
+
+        // Pouch capture riding.
+        if (event.getPlayer().isPassenger() && event.getItemStack().getItem().equals(ItemAlias.POUCH))
+        {
+            LivingEntity entity = (LivingEntity) event.getPlayer().getVehicle();
+
+            // Check for null.
+            if (entity == null) return;
+
+            // Pouch capture entity.
+            for (ItemContent i : horses)
+            {
+                if (i.getHorse().equals(entity.getType()))
+                {
+                    InteractionResult result = capture(i, (AbstractHorse)entity, event.getPlayer(), event.getHand());
+
+                    if (result.equals(InteractionResult.SUCCESS))
+                    {
+                        event.setCancellationResult(InteractionResult.SUCCESS);
+                        event.setCanceled(true);
+                    }
+                }
+            }
+        }
+    }
+
+    @SubscribeEvent
+    public static void eventInteract(PlayerInteractEvent.EntityInteract event)
+    {
+        if (event.getItemStack().getItem().equals(ItemAlias.POUCH))
+        {
+            List<ItemContent> horses = Arrays.asList(ItemAlias.POUCH_WITH_HORSE, ItemAlias.POUCH_WITH_SKELETON_HORSE, ItemAlias.POUCH_WITH_ZOMBIE_HORSE);
+
+            // Pouch capture entity.
+            for (ItemContent i : horses)
+            {
+                if (i.getHorse().equals(event.getTarget().getType()))
+                {
+                    InteractionResult result = capture(i, (AbstractHorse)event.getTarget(), event.getPlayer(), event.getHand());
+
+                    if (result.equals(InteractionResult.SUCCESS))
+                    {
+                        event.setCancellationResult(InteractionResult.SUCCESS);
+                        event.setCanceled(true);
+                    }
+                }
+            }
+        }
+    }
+
+    @SubscribeEvent
+    public static void eventBlock(PlayerInteractEvent.RightClickBlock event)
+    {
+        // Pouch release block.
+        if (event.getItemStack().getItem() instanceof ItemContent)
+        {
+            InteractionResult result = release((ItemContent) event.getItemStack().getItem(), event.getPlayer(), event.getHand(), event.getHitVec().getLocation());
+
+            if (result.equals(InteractionResult.SUCCESS))
+            {
+                event.setCancellationResult(InteractionResult.SUCCESS);
+                event.setCanceled(true);
+            }
+        }
+    }
+
+    /* Capture Feature */
+
+    public static InteractionResult capture(ItemContent item, AbstractHorse horse, Player player, InteractionHand hand)
+    {
+        ItemStack stack = player.getItemInHand(hand);
+
+        // Check for cooldown.
+        if (player.getCooldowns().isOnCooldown(stack.getItem())) return InteractionResult.FAIL;
+
+        // Create pouch.
+        ItemStack pouch = new ItemStack(item);
+        if (pouch.getTag() == null) pouch.setTag(new CompoundTag());
+
+        // Content compound.
+        CompoundTag content = new CompoundTag();
+        horse.save(content);
+
+        // Data compound.
+        CompoundTag data = item.getData(horse);
+        System.out.println(data);
+
+        // Add custom name.
+        if (horse.hasCustomName()) pouch.setHoverName(horse.getCustomName());
+
+        // Add data.
+        pouch.getTag().put("Content", content);
+        pouch.getTag().put("Data", data);
+
+        // Add data from previous stack.
+        addData(stack, pouch);
+
+        // Set cooldown.
+        addCooldown(player);
+
+        // Particles and sound.
+        addParticles(player, horse);
+
+        // Remove entity.
+        horse.discard();
+
+        // Set item to stack.
+        player.setItemInHand(hand, pouch);
+
+        return InteractionResult.SUCCESS;
+    }
+
+    /* Release Feature */
+
+    public static InteractionResult release(ItemContent item, Player player, InteractionHand hand, @Nullable Vec3 pos)
+    {
+        AbstractHorse horse = (AbstractHorse) item.getHorse().create(player.level);
+        ItemStack stack = player.getItemInHand(hand);
+
+        // Check for cooldown and horse.
+        if (player.getCooldowns().isOnCooldown(stack.getItem()) || horse == null) return InteractionResult.FAIL;
+
         // Add data from compound.
         horse.readAdditionalSaveData(player.getItemInHand(hand).getOrCreateTag().getCompound("Content"));
 
-        // Add pos and rot mimicking player.
-        horse.setPos(pos);
+        // Add pos
+        if (pos != null) horse.setPos(pos);
+        else horse.setPos(player.getPosition(0.0F));
+
+        // Add rot.
         horse.setYHeadRot(player.getYHeadRot());
         horse.setYBodyRot(player.getYHeadRot());
         horse.setYRot(player.getYRot());
@@ -47,44 +201,15 @@ public class UtilPouch
         // Particles and sound.
         addParticles(player, horse);
 
-        if (ride) player.startRiding(horse);
+        if (pos == null) player.startRiding(horse);
 
-        return pouch;
+        // Set item to stack.
+        player.setItemInHand(hand, pouch);
+
+        return InteractionResult.SUCCESS;
     }
 
-    public static ItemStack capture(AbstractHorse horse, Player player, InteractionHand hand)
-    {
-        // Remove leash.
-        if (horse.isLeashed()) horse.dropLeash(true, true);
-
-        // Create Horse Pouch.
-        ItemStack pouch = new ItemStack(ItemAlias.POUCH_WITH_HORSE);
-        pouch.setTag(new CompoundTag());
-
-        // Add custom name.
-        if (horse.hasCustomName()) pouch.setHoverName(horse.getCustomName());
-
-        // Save horse data to compound.
-        CompoundTag content = new CompoundTag();
-        horse.save(content);
-
-        // Add content
-        pouch.getOrCreateTag().put("Content", content);
-
-        // Add data.
-        addData(player.getItemInHand(hand), pouch);
-
-        // Set cooldown.
-        addCooldown(player);
-
-        // Particles and sound.
-        addParticles(player, horse);
-
-        // Remove entity.
-        horse.discard();
-
-        return pouch;
-    }
+    /* Tools */
 
     private static void addParticles(Player player, AbstractHorse horse)
     {
@@ -95,7 +220,13 @@ public class UtilPouch
             player.playSound(SoundEvents.ENDER_EYE_DEATH, 1.0F, 1.0F);
 
             // Poof
-            UtilParticle.spawnPoofParticles(horse);
+            for(int i = 0; i < 20; ++i)
+            {
+                double d0 = horse.getRandom().nextGaussian() * 0.02D;
+                double d1 = horse.getRandom().nextGaussian() * 0.02D;
+                double d2 = horse.getRandom().nextGaussian() * 0.02D;
+                horse.level.addParticle(ParticleTypes.POOF, horse.getRandomX(1.0D), horse.getRandomY(), horse.getRandomZ(1.0D), d0, d1, d2);
+            }
         }
     }